Prevent double renames by Photo-Renamer.py
It just seems wrong to rename a file because the filename has no date
but the folder name does, only to rename the file again because the
EXIF date is different.

So this update makes sure that a file date is captured by either the
filename or the parent folder name and then that name is used unless an
EXIF date is present.  The EXIF date always wins.