Photo-Backup: Update similar backups
When the photo to be backed up has a higher resolution or is the only
copy with a valid Date Taken, the backup is now automatically replaced.

Photos with a lower resolution are automatically removed.

That means that most of the time, no photos will still need to be
processed after running a backup.