PyBackup.py: A simple file backup utility
PyBackup reads ~\Backup.ini which contains a comma separated list of
folders. The first name in each line is the source folder and the
second line is the destination folder.

It then backs up every folder in that list.  To expedite subsequent
backups a crc.csv file is created in every source and destination
folder.  When the CRC values listed in those files agree, those files
are not backed up.  If a file changes it will be backed up and the CRC
for that file updated.