PyBackup.py: use filenames in hash files
Using a pathname in a hash file may cause unnecessary work when the
path used to access a file is different, but the file and CRC is the
same.  So strip pathnames and save filenames.  By doing both, old CRC
files do not even need to change in order to work with this improvement.

Print a bit more with each folder so that I can see how well the backup
is progressing.  Or more to the point, so I can see if it is doing too
much work.