PyBackup: add verification, improve error handling
The new -v switch can be used to verify that all data files match the
current CRC control files.

When a CRC control file is corrupt, it will be deleted and ignored.

A log file is now created each time the program runs. It can be used to
review the details of the exceptions that occurred during either a
backup or verify operation.

Make sure that a CRC is not added to the destination control file
unless that file exists. Ensure that the destination CRC matches the
file found at that location whenever a new CRC is actually computed.

Backups can potentially be faster now because the destination CRC will
never be computed twice.