PyBackup: use stats dict
Instead of using a set of counters, use another dictionary of counters
much like how errors has already been used. Logging has been updated to
make it easy to manage both errors and non-error counters.

I guess I could sum the stats for a running counter.

recursive_mkdir had a bug in it that could be reproduced when a path to
a non-existent drive was used.

I also fixed a bug with regards to how the ini file was processes. The
pathnames are now quoted and syntax errors are simply reported as
errors without causing any exception.